Test saving and credits

+++ Assets/Scripts/Player/PlayerDashing.cs
@@ -61,7 +61,7 @@ public class PlayerDashing : MonoBehaviour {
             //Actually Move the player
             controller.Move(MoveDirect);
 
-            if (gameObject.GetComponent<PlayerEquipment>().equippedBoot == (int)boot.Trailblazer)
+            if ((int)gameObject.GetComponent<PlayerEquipment>().equippedBoot == (int)boot.Trailblazer)
             {
                 trailBlazerDropTimer += Time.deltaTime;
                 if (trailBlazerDropTimer > 0.03f)
@@ -81,7 +81,7 @@ public class PlayerDashing : MonoBehaviour {
             dashTimeRemaining = dashDuration;
             trailBlazerDropTimer = 0.1f;
 
-            if (gameObject.GetComponent<PlayerEquipment>().equippedBoot == (int)boot.Trailblazer)
+            if ((int)gameObject.GetComponent<PlayerEquipment>().equippedBoot == (int)boot.Trailblazer)
             {
               //  Instantiate(TrailBlazerExplosion, transform.position, new Quaternion(0, 0, 0, 0));
